pyproject.toml
[project]
name = "openclaw-like-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "flyte>=2.0.9",
    "fpdf2>=2.8.7",
    "gradio>=6.9.0",
    "lxml>=6.0.2",
    "openai>=2.28.0",
    "openrouter>=0.7.11",
    "pandoc>=2.4",
    "python-pptx>=1.0.2",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "isort>=8.0.1",
    "ruff>=0.15.6",
]
